Add prop validation to Vehicle component
