+++ src/App.js
@@ -12,8 +12,8 @@ import Login from './components/Login';
 import Register from './components/Register';
 import Home from './components/Home';
 import Profile from './components/Profile';
+import Vehicles from './components/Vehicles';
 import Vehicle from './components/Vehicle';
-import Detail from './components/Detail';
 
 import { logout } from './actions/auth';
 import { clearMessage } from './actions/message';
@@ -91,8 +91,8 @@ const App = () => {
             <Route exact path="/login" component={Login} />
             <Route exact path="/register" component={Register} />
             <Route exact path="/profile" component={Profile} />
-            <Route exact path="/vehicle" component={Vehicle} />
-            <Route path="/vehicle/:id" component={Detail} />
+            <Route exact path="/vehicle" component={Vehicles} />
+            <Route path="/vehicle/:id" component={Vehicle} />
           </Switch>
         </div>
       </div>
